backend-development/starter-templates/task-tracker-api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
pytest==7.4.2
pytest-flask==1.3.0
black==23.9.1
flake8==6.1.0
gunicorn==21.2.0 